requirements.txt
python-dotenv==1.2.1
requests==2.32.5
beautifulsoup4==4.13.3
openai==1.100.2
google-genai==1.47.0
pymilvus[milvus-lite]==2.6.11
langfuse==3.7.0

